Update showAnimated: to included presented view controllers.
I was having problems with the alert not showing up on presented modals because it was only being added to view controllers in the root navigation controllers stack. This doesn't include the controllers that were presented with a modal. This fixes that problem by checking for a presented view controller and reasigning that to the view controller that the alert is added to.

Also made the bakcground a little bit darker.